Propagate register size to mask, offsets, fields
We're using this script in the imxrt-rs project. PWM registers on
iMXRT processors are 16 bits. But, this script is always generating
masks, offsets, and enumerated values that are hard-coded to u32s.
When we try bit operations -- like those inlined by the read_reg!,
write_reg!, or modify_reg! -- the compiler yells at us. We're trying
to AND, OR, and shift u32 values with u16 registers.

The masks, offsets, and enumerated values for a field should be based
on the overarching register. This commit updates the RAL generation
script to propagate the register size to the nodes that generate the
corresponding Rust code. We should see that the constants associated
with PWM fields are now correctly represented as u16s in our project.

This diff should have no noticeably effects on this repository,
particularly if all registers are 32 bits. But, it might be useful for
any other future users.