Ensure alignment when computing reserved registers
If the routine for computing reserved peripheral memory places at least
one u32 after a u16 register, the compiler could insert an extra u16 of
padding to align the u32 reservation. The extra padding is not
considered by the routine. This can result in incorrect register offsets
for all registers placed after the reservation.

This commit only uses byte arrays for register reservations. Byte
arrays, requiring one byte alignment, are the simplest way to reserve
peripheral memory. This will result in a larger diff when re-generating
the RAL.